Refactored Folder Structure, Removed unused code (firstsection/Card), Added alt text to Card.jsx img tag, Updated imports and exports.
