--- src/components/CardSecondSection/index.js
@@ -1,2 +1,0 @@
-import CardSecondSection from "./CardSecondSection";
-export default CardSecondSection;
