+++ src/containers/CardList/CardList.module.scss
@@ -1,4 +1,4 @@
-.CardListSecondSection {
+.CardList {
     display: flex;
     flex-direction: column;
     align-items: center;
