+++ src/containers/CardList/index.js
@@ -1,0 +1,2 @@
+import CardList from "./CardList";
+export default CardList;
