--- src/containers/CardListSecondSection/index.js
@@ -1,2 +1,0 @@
-import CardListSecondSection from "./CardListSecondSection";
-export default CardListSecondSection;
